# backend/docs/swagger.yaml
definitions:
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handlers.taskRequest:
    properties:
      completed:
        type: integer
      description:
        type: string
      dueDate:
        type: string
      title:
        type: string
    required:
    - completed
    - description
    - dueDate
    - title
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: This endpoint logs in a user by username and password.
      parameters:
      - description: Login Information
        in: body
        name: LoginRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - Auth
  /logout:
    post:
      description: This endpoint logs out a user by clearing the JWT cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Logout user
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: This endpoint registers a new user with a username and password.
      parameters:
      - description: Register Information
        in: body
        name: RegisterRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register new user
      tags:
      - Auth
  /tasks:
    get:
      consumes:
      - application/json
      description: Get a list of tasks with pagination.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of tasks
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not logged in
        "500":
          description: Internal Server Error
      summary: List tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Add a new task to the database.
      parameters:
      - description: Create Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/handlers.taskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Task created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Failed to bind JSON
        "401":
          description: User not logged in
        "500":
          description: Internal Server Error
      summary: Create new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific task from the database.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task deleted successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not logged in
        "500":
          description: Internal Server Error
      summary: Delete task
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Get details of a specific task.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task details
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not logged in
        "500":
          description: Internal Server Error
      summary: Get task
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update details of a specific task.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/handlers.taskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Task updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Failed to bind JSON
        "401":
          description: User not logged in
        "500":
          description: Internal Server Error
      summary: Update task
      tags:
      - tasks
swagger: "2.0"